requirements.txt
chromadb
tiktoken
flask
langchain>=0.3
langchain-openai>=0.2
langgraph>=0.2
pydantic>=2.0
httpx
aiosqlite
joblib>=1.3
scikit-learn>=1.3
sqlalchemy>=2.0
pytest
pytest-asyncio
pytest-mock
python-dotenv
